Simplify code by removing unecessary "Droid_Generic" Class.
Realized that the program requirements were simpler than I initially thought. As result,
the Droid_Generic class had no real purpose in the program as-is, and was removed. All
variables/methods associated with it have been transfered up into the Abstract Droid class.

As far as the front end, program should behave identically to last commit.

NOTE: The program is currently set so that all features have basically the same cost.
Should the program need to be expanded in a way that adjusts the cost of each feature,
look back into the previous commit with the "Droid_Generic" class. It was essentially set
up so that each material/color/model selection had an associated cost, as well as
the start of the logic for "clearance" and "random price" (including cents) functionality.
Once setup in Droid_Generic, the same functionality should be able to be implimented
in inheriting classes fairly easily, creating semi-dynamic prices for all features.
